third_party/blink/web_tests/fast/html/imports/import-insert-expected.txt

CONSOLE WARNING: line 6: HTML Imports is deprecated and will be removed in M80, around February 2020. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
It shouldn't be loaded before insertion.
PASS toInsert.import is null
It should be loaded once it is inserted.
PASS toInsert.import.querySelector('h1').innerHTML is 'Hello'
PASS successfullyParsed is true

TEST COMPLETE
